Build the index a region at a time, from a dump or a database
Two point seven million systems fit in memory and two hundred million
do not, and every road into the index went through a structure sized
by the galaxy. This is the road that does not.

`galos_index::Build` is one type and one pass: `begin`, `push` a system
per record, `finish`. A pushed system is appended to the spill file of
its bucket, a fixed level-4 cell, so one file handle is open at a time
whatever shape the galaxy is; the counts are tallied while writing, so
the cut is formed from the buckets afterwards and nothing is counted
beforehand. A bucket over budget is divided by re-reading that bucket
alone, one level deeper, and a piece that cannot divide -- at
`MAX_LEVEL`, or holding systems that share a position -- is reported as
a region over budget rather than hidden. Then the crown settles the
cells above the cut, each region is built from its own spill minus what
the crown took, written, and dropped.

A region can be built alone because the coupling is bounded, and that
is the whole of why this works: a region's crown cells are its own
ancestors, `region.level` of them, each owning at most
`internal_slice`, so a region can lose only its brightest
`level x internal_slice` systems and it is enough to offer exactly
those. Nothing fainter can be claimed -- a candidate that finds no room
proves the crown was full, and the crown never empties.
`a_regional_build_is_the_whole_build` holds it to a batch build over
the same systems: same cells, same `rank_lo`, `rank_hi` and
`child_mask`, byte-equal payloads.

`GALOS_REGION_BUDGET` is the dial rather than the galaxy. Over 240,000
systems the event sink's live tree peaked at 389 MiB and the same input
through a region build at 82.8 MiB, and at a 50 k budget, 24 regions,
65.5 MiB. The whole seven-day slice -- 730,544 systems, 3,119,799
bodies -- builds in 64.9 s at 198.6 MiB, and the directory it writes is
identical in every figure to the one built from Postgres, which needs a
43-minute import first.

Two callers push into it and there is no third: `galos_db::index`
merges a pair of ordered SQL cursors, and `bin/sync/spansh.rs` reads
the dump once. `spansh` is a new crate for that dump and nothing else,
30,000 systems a second at 7.4 MB resident, one `serde_json` call per
line into a reused buffer.

A stopped build answers a value rather than an error: `Built::Stopped`
carrying what it abandoned, and up to the point of no return the
directory is left exactly as it was found -- the names table staged
under `.building/` and published by rename, where a cold build used to
overwrite the served table a chunk at a time as it read. A SIGINT 0.9 s
into a cold build now exits in 0.023 s.

`galos_journal` is dissolved, having been three things in a trenchcoat.
The reading of a live journal directory went to `elite_journal`, which
is the crate that knows what a journal line is; the galaxy accumulator
and the body store went to `galos_index`, which is what they were
always building; the sink went to the one program that has sinks.
`src/bin/` goes to `bin/` with it, so the crate root holds the library
the binaries share rather than the binaries themselves.

Four bugs came out of having two derivations to diff on real data,
which is what `tests/derivations_agree.rs` is for. Anarchy was not a
security reading, so every anarchy system's row was refused and its
bodies went with it behind the foreign key -- thirteen per cent of the
file. Two star records under one name double-counted a system's light,
the dump listing a star twice differing only in `bodyId` and
`distanceToArrival`. An uploader was filed as `unknown`, the relay
carrying only a commander across the channel, which is why all 277,551
body files differed between the two routes. And barycentres came back
in Postgres heap order for want of a secondary key, 7,680 files
differing by ordering alone. The fixture now carries a same-named twin
star, a second barycentre and a non-commander reporter -- the three
shapes whose absence let all of this pass.

`galos-index info` measured shard directories rather than payloads,
ever since the sharding landed, and `galos_map`'s perf guard was not
compiled for want of a `mod perf;`.

Tests provision their own database. `galos_db::testing::Scratch` hands
a test a fresh migrated copy of a template and drops it afterwards, and
a reaper clears any scratch database whose process is dead, which is
the panic path. `TEST_DATABASE_URL` names a server now, not a database,
so the suite needs nothing created by hand and leaves nothing behind.

+++ galos_index/src/aggregate.rs
@@ -1,25 +1,24 @@
 //! What a cell carries for its whole subtree, held so it composes exactly.
 //!
-//! A cell stands for every system beneath it, and it has to say something true
-//! about them whether or not their records are loaded. So it carries totals,
-//! the count, the flux, the two weighted centroids, the brightest magnitude,
-//! and those totals are the aggregate `T(c)`. With the payload absent the total
-//! is drawn as it stands; with the payload present the residual, the total less
-//! the moments of the slice that arrived, is drawn instead, so no system counts
-//! twice.
+//! A cell stands for every system beneath it, and says something true about
+//! them whether or not their records are loaded: the count, the flux, the two
+//! weighted centroids, the brightest magnitude — the aggregate `T(c)`. With
+//! the payload absent the total is drawn as it stands; with the payload
+//! present the residual, the total less the moments of the slice that
+//! arrived, is drawn instead, so no system counts twice.
 //!
-//! Everything additive is kept as a sum, because sums compose and the things
-//! read off them do not. Flux per temperature bucket, counts, and age buckets
-//! add. The centroids and spreads come out of [`Moments`], which keeps the
-//! moments they are read from rather than the answers. Two weightings run at
-//! once and diverge wherever the bright stars sit off centre: the glow follows
-//! the light and the density follows the count.
+//! Everything additive is kept as a sum, sums composing where the things read
+//! off them do not: flux per temperature bucket, counts, age buckets. The
+//! centroids and spreads come out of [`Moments`], which keeps the moments
+//! they are read from rather than the answers. Two weightings run at once and
+//! diverge wherever the bright stars sit off centre: the glow follows the
+//! light and the density follows the count.
 //!
-//! One field does not compose by adding, and does not need to. `m_min`, the
-//! brightest absolute magnitude in the subtree, composes by taking the smaller
-//! of two, which is why it is stored and not derived from the flux, a sum that
-//! has lost the single brightest star. It answers the photometric cull on the
-//! stored total, never on a residual, so [`Aggregate::remove`] leaves it be.
+//! `m_min`, the brightest absolute magnitude in the subtree, composes by
+//! taking the smaller of two rather than by adding, and is stored because a
+//! summed flux has lost the single brightest star. It answers the photometric
+//! cull on the stored total, never on a residual, so [`Aggregate::remove`]
+//! leaves it be.
 
 use crate::geometry::CellId;
 use crate::moments::Moments;
@@ -33,20 +32,16 @@ pub const TEMP_BUCKETS: usize = 6;
 /// Age buckets for the Recency axis, which a prefix sum answers any span from.
 pub const AGE_BUCKETS: usize = 8;
 
-/// The temperature range the buckets span, log-spaced between them.
-///
-/// The coolest star worth coloring and the hottest whose blue has stopped
-/// moving; [`temp_bucket`] bins the range and [`bucket_temperature`] names a
-/// point back out of a bucket.
+/// The temperature range the buckets span, log-spaced between them: the
+/// coolest star worth coloring and the hottest whose blue has stopped moving.
+/// [`temp_bucket`] bins the range and [`bucket_temperature`] names a point
+/// back out of a bucket.
 const TEMP_LO: f64 = 2000.0;
 const TEMP_HI: f64 = 50000.0;
 
 /// Which temperature bucket a star falls in, log-spaced across the stellar
-/// range and clamped at both ends.
-///
-/// The ends are the coolest star worth coloring and the hottest whose blue has
-/// stopped moving; between them the buckets are even in log temperature, which
-/// is where color is even.
+/// range and clamped at both ends. The buckets are even in log temperature,
+/// which is where color is even.
 pub fn temp_bucket(temperature_k: f64) -> usize {
     let t = temperature_k.clamp(TEMP_LO, TEMP_HI);
     let f = (t.ln() - TEMP_LO.ln()) / (TEMP_HI.ln() - TEMP_LO.ln());
@@ -69,8 +64,8 @@ pub fn bucket_temperature(bucket: usize) -> f64 {
 ///
 /// Built from single systems with [`of_system`](Self::of_system), rolled up
 /// with [`merge`](Self::merge), and drawn over its own loaded slice through
-/// [`remove`](Self::remove). Everything but `m_min` is a sum, so a set split any
-/// way and rejoined is the same aggregate.
+/// [`remove`](Self::remove). Everything but `m_min` is a sum, so a set split
+/// any way and rejoined is the same aggregate.
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub struct Aggregate {
     /// Brightest absolute magnitude in the subtree, the smallest number, or
@@ -82,7 +77,8 @@ pub struct Aggregate {
     flux: [f64; TEMP_BUCKETS],
     /// Position moments weighted by flux, for the glow's centroid and spread.
     light: Moments,
-    /// Position moments weighted by count, for the count-weighted centroid and extent.
+    /// Position moments weighted by count, for the count-weighted centroid
+    /// and extent.
     mass: Moments,
     /// Counts per age bucket, a column of the record so a Recency span can be
     /// answered by prefix sum off the aggregates alone. Every build writes it
@@ -90,15 +86,10 @@ pub struct Aggregate {
     /// the `updated_at` the payload carries, which is the same clock binned
     /// finer.
     ///
-    /// `u32` rather than `u64`, which halves what the column costs and loses
-    /// nothing: a bucket counts systems, the buckets of one cell sum to its
-    /// `count`, and the root's count is the galaxy's — 129 million against the
-    /// four billion a `u32` holds, so there are five doublings of headroom
-    /// over every system on record. Eight buckets at four bytes is thirty-two
-    /// bytes a cell. Halving that again by storing each bucket as a `u16`
-    /// share of `count` rounds the smallest buckets away to nothing over a
-    /// real build, and the smallest bucket is the recently-changed one the
-    /// axis exists to show. Exact at twice the width is the better trade.
+    /// `u32`, and exact: a bucket counts systems and the buckets of a cell
+    /// sum to its `count`, so four billion is ample over the galaxy's 129
+    /// million, where a `u16` share of `count` would round the smallest
+    /// bucket — the recently-changed one the axis exists to show — away.
     aged: [u32; AGE_BUCKETS],
 }
 
@@ -117,20 +108,21 @@ impl Aggregate {
     ///
     /// Its flux is `10^(-0.4*M)`, the linear form magnitudes sum in, dropped
     /// into the bucket its temperature falls in. Its position enters the glow
-    /// weighted by that flux and the density weighted by one, and its magnitude is
-    /// the brightest the aggregate has seen until something brighter merges in.
+    /// weighted by that flux and the density weighted by one, and its
+    /// magnitude is the brightest the aggregate has seen until something
+    /// brighter merges in.
     pub fn of_system(
         position: [f64; 3],
         absolute_magnitude: f64,
         temperature: f64,
-        age_bucket: usize,
+        age_bucket: u32,
     ) -> Aggregate {
         let f = Magnitude(absolute_magnitude).flux().0;
         let mut flux_by_bucket = [0.0; TEMP_BUCKETS];
         flux_by_bucket[temp_bucket(temperature)] = f;
         let mut aged = [0; AGE_BUCKETS];
-        if age_bucket < AGE_BUCKETS {
-            aged[age_bucket] = 1;
+        if (age_bucket as usize) < AGE_BUCKETS {
+            aged[age_bucket as usize] = 1;
         }
         Aggregate {
             m_min: Some(absolute_magnitude as f32),
@@ -170,7 +162,7 @@ impl Aggregate {
     /// The additive fields subtract exactly, being the inverse of
     /// [`merge`](Self::merge). `m_min` is left untouched: a residual splat is
     /// never culled on it, and the brightest single star cannot be recovered
-    /// from a flux that has already summed it away.
+    /// from a summed flux.
     pub fn remove(self, slice: Aggregate) -> Aggregate {
         let mut flux = self.flux;
         let mut aged = self.aged;
@@ -296,10 +288,10 @@ fn min_opt(a: Option<f32>, b: Option<f32>) -> Option<f32> {
 /// children it has, and the totals it stands for.
 ///
 /// A node at level `L` owns ranks `[rank_lo, rank_hi)` of its subtree's
-/// magnitude order, holding only what its ancestors did not, so drawing a node
-/// with its loaded ancestors is exactly the union with no system twice. The
-/// `aggregate` is the total over the whole subtree, not the slice; with the
-/// slice absent it is drawn as it stands, and with the slice present the
+/// magnitude order, holding only what its ancestors did not, so drawing a
+/// node with its loaded ancestors is exactly the union with no system twice.
+/// The `aggregate` is the total over the whole subtree, not the slice; with
+/// the slice absent it is drawn as it stands, and with the slice present the
 /// residual is drawn instead.
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub struct Cell {
@@ -413,7 +405,7 @@ mod tests {
     }
 
     /// The two centroids diverge: a bright star and a dim one meet in the
-    /// middle by count, but the glow leans hard toward the bright one.
+    /// middle by count, where the glow leans toward the bright one.
     #[test]
     fn the_glow_and_the_map_centre_differ() {
         let bright = Aggregate::of_system([0.0, 0.0, 0.0], -1.0, 15000.0, 0);
@@ -430,7 +422,7 @@ mod tests {
     #[test]
     fn a_split_conserves_the_subtree() {
         let systems = [
-            ([10.0, 0.0, 0.0], 3.0, 6000.0, 1usize),
+            ([10.0, 0.0, 0.0], 3.0, 6000.0, 1u32),
             ([0.0, 10.0, 0.0], 7.0, 3500.0, 2),
             ([0.0, 0.0, 10.0], -1.0, 20000.0, 0),
             ([-5.0, -5.0, -5.0], 5.0, 4800.0, 3),
@@ -469,14 +461,14 @@ mod tests {
     #[test]
     fn remove_leaves_the_residual() {
         let slice: Aggregate = [
-            ([1.0, 0.0, 0.0], 2.0, 6000.0, 0usize),
+            ([1.0, 0.0, 0.0], 2.0, 6000.0, 0u32),
             ([0.0, 1.0, 0.0], 4.0, 4000.0, 1),
         ]
         .iter()
         .map(|&(p, m, t, a)| Aggregate::of_system(p, m, t, a))
         .collect();
         let rest: Aggregate = [
-            ([5.0, 5.0, 5.0], 6.0, 3500.0, 2usize),
+            ([5.0, 5.0, 5.0], 6.0, 3500.0, 2u32),
             ([-2.0, 3.0, 1.0], 8.0, 3200.0, 3),
             ([0.0, 0.0, 9.0], 1.0, 12000.0, 0),
         ]
