+++ galos_index/src/derive.rs
@@ -1,22 +1,18 @@
 //! The rules both derivations of the index have to answer alike.
 //!
-//! A directory is produced two ways. `galos_db::index` derives one from
-//! Postgres, over every system anybody has ever reported; `galos_journal`
-//! derives one from a single pilot's own journal files, with no database
-//! anywhere. They write the same format, for the same client, and a running
-//! `galos-sync` maintains one directory from live events while catching it up
-//! from the database, so the two answers land in the same files. Where they
-//! read the same fact they have to read it the same way: a system binned into
-//! Recency bucket three out of the database and bucket four out of a journal
-//! is one galaxy disagreeing with itself about when it was last heard from,
-//! and the disagreement shows as a cell whose counts do not add up to the
+//! A directory is produced two ways. [`crate::cold`] raises one from
+//! records, which `galos_db::index` reads out of Postgres over every system
+//! anybody has ever reported; [`crate::galaxy`] accumulates one from events,
+//! which reach it off the wire or out of a single pilot's own journal files
+//! with no database anywhere. They write the same format, for the same
+//! client, and a running `galos-sync` maintains one directory from both at
+//! once. Where they read the same fact they must read it the same way: a
+//! system binned into Recency bucket three out of the database and bucket
+//! four out of a journal shows as a cell whose counts do not add up to the
 //! stars drawn inside it.
 //!
-//! Each rule below stood written out in both crates, word for word in places,
-//! with a comment in each admitting the other existed. They live here
-//! instead. This crate is the format both sides agree on and both already
-//! link it, so a rule kept once cannot drift; a rule kept twice only has not
-//! drifted yet.
+//! So each such rule is kept here once, this crate being the format both
+//! sides agree on.
 
 use crate::meta::SystemBodies;
 use chrono::NaiveDateTime;
@@ -26,11 +22,9 @@ use galos_photometry::{ClassLight, Flux, Magnitude};
 /// last written. Updated today lands in bucket 0, untouched for a decade in
 /// bucket 7.
 ///
-/// Replaces the identical arrays in `galos_db::index` and
-/// `galos_journal::galaxy`. The count of buckets is part of the published
-/// format — a cell aggregate is a count per bucket, and the client's Recency
-/// control indexes straight into them — so the edges are not a builder's
-/// choice to make.
+/// The count of buckets is part of the published format — a cell aggregate
+/// is a count per bucket, and the client's Recency control indexes straight
+/// into them — so the edges are not a builder's choice to make.
 pub const AGE_EDGES: [i64; 7] = [1, 7, 30, 90, 365, 1095, 3650];
 
 /// Which Recency bucket an age in days falls in, `0..8`.
@@ -39,29 +33,27 @@ pub const AGE_EDGES: [i64; 7] = [1, 7, 30, 90, 365, 1095, 3650];
 /// an age past the last edge saturates at 7 rather than running off the end.
 /// A negative age — a report stamped in the future, which a client can write
 /// — falls in bucket 0 with everything else fresh.
-pub fn age_bucket(days: i64) -> usize {
-    AGE_EDGES.iter().filter(|&&edge| days >= edge).count()
+///
+/// A `u32` and not the `usize` an array index wants, because where it is kept
+/// is a [`System`](crate::System), and that record is written to disk as its
+/// own bytes; see its doc.
+pub fn age_bucket(days: i64) -> u32 {
+    AGE_EDGES.iter().filter(|&&edge| days >= edge).count() as u32
 }
 
 /// How lately a system was updated, in the two forms the index wants it: the
 /// Recency bucket the cell aggregates count by and the Unix second the
 /// payload carries.
 ///
-/// One reading of `at`, so the two cannot disagree about a system. The bucket
-/// alone was what the index carried and the buckets are days wide, which
-/// answers a Recency span of thirty days and none of the five spans shorter
-/// than a day — the end of the control the map is actually used at. So the
-/// second goes on the payload point beside it.
+/// One reading of `at`, so the two cannot disagree about a system. The
+/// buckets are days wide, where the map's Recency control goes down to a
+/// minute, which is what the second beside them answers.
 ///
-/// `u32`, which is Unix seconds to 2106 and four bytes rather than eight on a
-/// record of thirty-five. Clamped rather than wrapped: `at` comes off a
-/// journal entry and a client can write whatever it likes there, and a year
-/// outside the `u32` range should read as the far end of the axis rather than
-/// fold back into the middle of it.
-///
-/// Replaces `galos_db::index::updated` and the same pair worked out inline in
-/// `galos_journal::galaxy`, where the clamp was written a second time.
-pub fn updated(at: NaiveDateTime, now: NaiveDateTime) -> (usize, u32) {
+/// `u32`: Unix seconds to 2106, four bytes rather than eight. Clamped rather
+/// than wrapped — `at` comes off a journal entry, which a client writes, so
+/// a year outside the range reads as the far end of the axis instead of
+/// folding into the middle of it.
+pub fn updated(at: NaiveDateTime, now: NaiveDateTime) -> (u32, u32) {
     (
         age_bucket((now - at).num_days()),
         at.and_utc().timestamp().clamp(0, u32::MAX as i64) as u32,
@@ -71,26 +63,18 @@ pub fn updated(at: NaiveDateTime, now: NaiveDateTime) -> (usize, u32) {
 /// One system's photometry, by the fallback chain: its scanned stars if it
 /// has any, else the class it is named for, else the default M dwarf.
 ///
-/// `stars` is the `(visual absolute magnitude, temperature)` of every scanned
-/// star — visual, the caller having already carried Elite's bolometric
-/// scanned figure through [`Magnitude::visual`], which is where a white dwarf
-/// keeps its faint brightness and a black hole falls to nothing. Their light
-/// adds, so the magnitudes combine to the one figure the sky is ordered by,
-/// and the tint is the brightest star's, that being the one that dominates
-/// what the pair looks like. With no stars `fallback_class` stands in, and
-/// with no class [`ClassLight::of`] hands back the M dwarf the galaxy is
-/// mostly made of.
+/// `stars` is the `(visual absolute magnitude, temperature)` of every
+/// scanned star — visual, the caller having already carried Elite's
+/// bolometric figure through [`Magnitude::visual`], which is where a white
+/// dwarf keeps its faint brightness and a black hole falls to nothing. Their
+/// light adds, so the magnitudes combine to the one figure the sky is
+/// ordered by, and the tint is the brightest star's, that being the one that
+/// dominates what the pair looks like. With no stars `fallback_class` stands
+/// in, and with no class [`ClassLight::of`] hands back the M dwarf the
+/// galaxy is mostly made of.
 ///
-/// Replaces the fold in `galos_db::index::system_input` and the identical one
-/// in `galos_journal::galaxy::Galaxy::system`. The two disagreeing would show
-/// as a system changing brightness when the same directory was next written
-/// by the other side.
-///
-/// Folded in one pass rather than collecting, since the database side runs
-/// this over a hundred and twenty-nine million systems: the flux sums as it
-/// goes and the brightest star is carried along, which is
-/// [`Magnitude::combine`] and a `min_by` without the vector in between. Ties
-/// keep the first star, as `min_by` did.
+/// Folded in one pass, no vector in between: the flux sums as it goes and
+/// the brightest star is carried along. A tie keeps the first star.
 pub fn lit(
     stars: impl Iterator<Item = (f64, f64)>,
     fallback_class: &str,
@@ -123,21 +107,14 @@ pub fn lit(
 ///
 /// Nearest the arrival point, ties broken by body id. The tie-break is what
 /// makes it an answer rather than a coin toss: a close pair is recorded at
-/// the same distance from arrival to the resolution the journal prints, and
-/// whichever of them the caller happened to read first would otherwise decide
-/// whether the system supercharges a drive.
+/// one distance from arrival to the resolution the journal prints.
 ///
 /// [`None`] for a system with no star on record, which leaves the caller to
-/// its own fallback — the database side coalesces to the `primary_star_class`
-/// column and the journal side to the class a plotted route named, those
-/// being statements about the same star from the only sources each has.
-///
-/// Replaces `galos_journal::galaxy::Galaxy::arrival_class` and the
-/// `DISTINCT ON (system_address) ... ORDER BY system_address,
-/// distance_from_arrival_ls, id` in `galos_db::index::metadata`, which is the
-/// same rule said in SQL. The boost table is written by both and patched by
-/// one, so a directory whose full build and whose watch passes ordered
-/// differently would disagree with itself about the same galaxy.
+/// its own fallback — the database side coalesces to the
+/// `primary_star_class` column and the journal side to the class a plotted
+/// route named. The same rule stands in SQL as the `DISTINCT ON
+/// (system_address) ... ORDER BY system_address, distance_from_arrival_ls,
+/// id` of `galos_db::index::metadata`.
 pub fn arrival_class(bodies: &SystemBodies) -> Option<&str> {
     bodies
         .stars
@@ -188,8 +165,7 @@ mod tests {
     /// past the last edge saturates rather than running off the end.
     ///
     /// The eight buckets are the published format: a cell aggregate is a
-    /// count per bucket and the client indexes straight into them, so this is
-    /// the table both derivations answer by.
+    /// count per bucket and the client indexes straight into them.
     #[test]
     fn the_edges_bin_where_the_format_says() {
         assert_eq!(age_bucket(-7), 0, "a report stamped in the future");
@@ -206,16 +182,19 @@ mod tests {
         assert_eq!(age_bucket(1095), 6);
         assert_eq!(age_bucket(3649), 6);
         assert_eq!(age_bucket(3650), 7);
-        assert_eq!(age_bucket(i64::MAX), AGE_EDGES.len(), "saturates at 7");
+        assert_eq!(
+            age_bucket(i64::MAX),
+            AGE_EDGES.len() as u32,
+            "saturates at 7"
+        );
     }
 
     /// The second the payload carries saturates at either end of the `u32`
     /// rather than wrapping into the middle of the axis.
     ///
     /// `updated_at` comes off a journal entry, which a client writes, so a
-    /// date in 1969 or in 2200 is a thing the builder will be handed. Wrapped,
-    /// either would land beside genuinely fresh systems and colour a cell by
-    /// a recency nothing in it has.
+    /// date in 1969 or in 2200 does reach the builder. Wrapped, either would
+    /// land beside genuinely fresh systems.
     #[test]
     fn the_clamp_saturates_rather_than_wrapping() {
         let now = at(1_700_000_000);
@@ -236,8 +215,7 @@ mod tests {
     /// With no stars the photometry is the class's, and with no class the
     /// default the galaxy is mostly made of.
     ///
-    /// The unscanned case is nearly every system there is, so the fallback is
-    /// the common path rather than the edge one.
+    /// The unscanned case is nearly every system there is.
     #[test]
     fn lit_falls_back_with_nothing_scanned() {
         let class = ClassLight::of("G");
@@ -256,10 +234,9 @@ mod tests {
     /// Scanned stars combine to one figure brighter than any of them, and the
     /// tint is the brightest star's.
     ///
-    /// A magnitude is inverted and logarithmic, so the two mistakes available
-    /// here are taking the dimmest star's tint and averaging light that
-    /// should add. Both are checked: the pair must come out brighter than its
-    /// brightest member, and wear that member's temperature.
+    /// A magnitude is inverted and logarithmic: the pair must come out
+    /// brighter than its brightest member, and wear that member's
+    /// temperature.
     #[test]
     fn lit_adds_the_light_and_takes_the_brightest_tint() {
         let (magnitude, temperature) =
@@ -283,8 +260,7 @@ mod tests {
     ///
     /// A close pair is recorded at one distance from arrival to the
     /// resolution the journal prints, so without the tie-break the answer
-    /// would be whichever star the caller read first — and the database side
-    /// reads them in whatever order a query returns.
+    /// would be whichever star the caller read first.
     #[test]
     fn arrival_class_breaks_a_tie_by_id() {
         let mut bodies = SystemBodies::default();
